Changes 16/04
Added Images and items to the DB can are on the listing page. Wont show on the index page due to that function is looking for items with the category "Featured" and haven't added any yet.

Listing page now has Plants and Toys categories with a limit of 12 just works on the screen size I'm using but we can change the hardcode in the function to best match the look on the day think  8 works on my laptop but not 100% sure.

Removed the blocks on repeated code on the index and listing page as he want us to not repeat code so I've added a Functions.php file that we can use if you guys want to add some more custom functions use that file.

Fixed the navbar so the logo doesn't spill on to the rest of the page anymore and noticed an issue where it wasn't fully at the top of the page so that's also been fixed.

Changed the layout of the index page it now shows "How It Works" below the image we have for the background of the index page and then Featured Listings is below that and have a image path and text in the DB thats being pulled to the page. ( Will change this to pull the actual image file once we get the CRUD HTML files)

Began creating a function to add emails to a table in the DB for the newsletter we have in the footer thought it would be a nice little feature so Ill get that finished off over the next few days or if anyone wants to finish it be my guest.