Changed 17/04
- Created a dropdown menu for the navbar. Allows the user to select different categories only or an option to select All products. (Will add more)

- Began trying to figure out the dynamic product page. Keep getting an error in which the product ID cant be found. need to look into this and figure it out

Changed some of the file structure so it doesn't look as clutter and more organized. Header and Footer are not in Private / Templates. Will do more just takes a while as I need to redo links and ensure everything is properly being access without any issues.
(Want to have all the productPages set in their own folder or change it to use dynamic pages. Would just require metadate in the URL to figure out the category. Once I figure out the items stuff I should be able to convert the code to do this also.

- Couple minor tweaks were also made.