Changes 21/04
Added comments to functions.

Implemented redirect "See More" button to Listings.php

Implemented Featured Listings function for index page and changed to DB for it.

Added More Products manually 👎 have to create CRUD tomorrow as it has yet to be uploaded

Changed Titles to each page

Changed redirect function was going to one for each page but realized I can just make it take a parameter for the page instead

Still good bit to work on but getting there